requirements.txt
dash==1.11.0
dash-core-components==1.9.1
dash-html-components==1.0.3
flake8==3.7.9 # dev requirement, but not bothering with separate requirements files
gunicorn==20.0.4
requests==2.23.0
pandas==1.0.3
pytest==5.4.1 # dev requirement, but not bothering with separate requirements files
psycopg2-binary==2.8.5
plotly==4.6.0
SQLAlchemy==1.3.16
